Move version validation to cmd.ts
